🚚 Rename folder to userId since we're not using CPF as identifier anymore
